Unit tests for the home functionality when user is authenticated have been added.
The tests consist of checking that all the posts that appear on the home are from users you follow, so the Following button must appear, or they are yours so no follow button appears. It is also verified that if you stop following a user, when you reload the page, all the posts from that user disappear.

The tests do not pass because there may be posts from users whom you stop following. This functionality needs to be added to Frontend.

The Follow/Following buttons take a while to update, they should appear directly with Following, and instead they start at Follow and after a few seconds change to Following.